(hds-1467) Moved components to own folder
