+++ packages/react/src/components/login/components/LoginContext.tsx
@@ -1,9 +1,9 @@
 import React, { createContext, useMemo } from 'react';
 
-import { OidcClientProps } from './client/index';
-import { createOidcClient } from './client/oidcClient';
-import { createBeacon, Beacon, ConnectedModule, SignalNamespace } from './beacon/beacon';
-import { emitInitializationSignals } from './beacon/signals';
+import { OidcClientProps } from '../client/index';
+import { createOidcClient } from '../client/oidcClient';
+import { createBeacon, Beacon, ConnectedModule, SignalNamespace } from '../beacon/beacon';
+import { emitInitializationSignals } from '../beacon/signals';
 
 type ContextProps = {
   children: React.ReactNode | React.ReactNode[] | null;
